+++ packages/client/src/types.ts
@@ -2,6 +2,7 @@ import { BlockInfo } from "@polkadot-api/observable-client"
 import { BlockHeader, HexString } from "@polkadot-api/substrate-bindings"
 import { ChainSpecData } from "@polkadot-api/substrate-client"
 import { Observable } from "rxjs"
+import { ArgsValueCompatHelper, CompatHelper } from "./compatibility"
 import { ConstantEntry } from "./constants"
 import {
   ApisFromDef,
@@ -16,14 +17,13 @@ import { EvClient } from "./event"
 import { RuntimeCall } from "./runtime-call"
 import { StorageEntry } from "./storage"
 import type {
-  TxEntry,
   OfflineTxEntry,
   TxBroadcastEvent,
+  TxEntry,
   TxFinalizedPayload,
   TxFromBinary,
 } from "./tx"
 import { ViewFn } from "./viewFns"
-import { ArgsValueCompatHelper, CompatHelper } from "./compatibility"
 
 export type { ChainSpecData }
 
@@ -115,7 +115,7 @@ export type OfflineConstApi<A extends Record<string, Record<string, any>>> = {
 
 export type UnsafeElement<E> = Record<string, Record<string, E>>
 
-export type { CompatHelper, ArgsValueCompatHelper }
+export type { ArgsValueCompatHelper, CompatHelper }
 export type CompatHelperApi<A extends Record<string, Record<string, any>>> = {
   [K in keyof A]: {
     [KK in keyof A[K]]: A[K][KK] extends {} | undefined
@@ -340,7 +340,7 @@ export interface PolkadotClient {
    *                     the tx.
    */
   submit: (
-    transaction: HexString,
+    transaction: Uint8Array,
     at?: HexString,
   ) => Promise<TxFinalizedPayload>
   /**
@@ -355,7 +355,7 @@ export interface PolkadotClient {
    *                     the tx.
    */
   submitAndWatch: (
-    transaction: HexString,
+    transaction: Uint8Array,
     at?: HexString,
   ) => Observable<TxBroadcastEvent>